+++ lib/modules/Cache/DurableObject.ts
@@ -1,6 +1,6 @@
 import {type TriFrostCFDurableObjectNamespace} from '../../types/providers';
 import {type LazyInitFn} from '../../utils/Lazy';
-import {DurableObjectStore} from '../_stores/DurableObject';
+import {DurableObjectStore} from '../_storage/DurableObject';
 import {TriFrostCache} from './_Cache';
 
 export class DurableObjectCache<Env extends Record<string, any> = Record<string, any>> extends TriFrostCache<Env> {