+++ lib/modules/Cache/KV.ts
@@ -1,6 +1,6 @@
 import {type TriFrostCFKVNamespace} from '../../types/providers';
 import {type LazyInitFn} from '../../utils/Lazy';
-import {KVStore} from '../_stores/KV';
+import {KVStore} from '../_storage/KV';
 import {TriFrostCache} from './_Cache';
 
 export class KVCache<Env extends Record<string, any> = Record<string, any>> extends TriFrostCache<Env> {