+++ lib/modules/RateLimit/KV.ts
@@ -1,6 +1,6 @@
 import {type TriFrostCFKVNamespace} from '../../types/providers';
 import {type LazyInitFn} from '../../utils/Lazy';
-import {KVStore} from '../_stores/KV';
+import {KVStore} from '../_storage/KV';
 import {
     TriFrostRateLimit,
     type TriFrostRateLimitStrategy,