+++ CHANGELOG.md
@@ -9,8 +9,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - **misc**: Migrate existing tests using node assertion to vitest
 - **misc**: Add test runs to CI for both node (20, 22) and bun runtime jobs
 - **misc**: Add coverage reporting to codecov to CI
+- **misc**: Add tests for KV, Redis storage modules
 - **deps**: vitest (dev dependency)
 - **deps**: @vitest/coverage-v8 (dev dependency)
+
+### Fixed
+- Fixed an issue in KV storage module where set would not work due to a conflicting object vs array conditional check
+- Fixed an issue in Redis storage module where set would not work due to a conflicting object vs array conditional check
 
 ### Removed
 - **deps**: nyc (as no longer in use)